# specs/001-oci-registry-server/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: jreg - OCI Distribution Spec v1.1.1
  description: |
    OCI-compliant container registry server implementing the OCI Distribution Specification v1.1.1.
    
    This API enables pushing and pulling OCI images and artifacts using Docker CLI, ORAS, and other OCI-compliant tools.
    
    **Key Features:**
    - Full OCI Distribution Spec v1.1.1 compliance
    - Content-addressable storage (SHA-256/SHA-512)
    - Chunked blob uploads with resume support
    - Multi-architecture image support (OCI Image Index)
    - Referrers API for artifact linking (signatures, SBOMs)
    - Tag management and discovery
    
    **Authentication:** Deferred for MVP (all endpoints publicly accessible)
  version: 1.0.0
  contact:
    name: jreg Development Team
  license:
    name: MIT
servers:
  - url: http://localhost:8080/v2
    description: Local development server
  - url: https://registry.example.com/v2
    description: Production server
...
paths:
...
  /{name}/blobs/{digest}:
    parameters:
      - $ref: '#/components/parameters/RepositoryName'
      - $ref: '#/components/parameters/Digest'
    
    head:
      operationId: checkBlobExists
      tags: [blobs]
      summary: Check if blob exists
      description: |
        Check if a blob exists in the repository (OCI Distribution Spec end-2).
        Returns 200 if blob exists, 404 if not found.
      responses:
        '200':
          description: Blob exists
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
              description: Blob size in bytes
            Docker-Content-Digest:
              schema:
                type: string
              description: Canonical digest of the blob
            Content-Type:
              schema:
                type: string
                example: application/octet-stream
        '404':
          $ref: '#/components/responses/BlobNotFound'
    
    # end-3: Blob download
    get:
      operationId: getBlob
      tags: [blobs]
      summary: Download blob
      description: |
        Download a blob from the repository (OCI Distribution Spec end-3).
        Supports HTTP Range requests for partial downloads.
      parameters:
        - name: Range
          in: header
          required: false
          schema:
            type: string
            example: bytes=0-1048575
          description: Request partial content (RFC 7233)
      responses:
        '200':
          description: Blob content
          headers:
            Content-Length:
              schema:
                type: integer
                format: int64
            Docker-Content-Digest:
              schema:
                type: string
            Content-Type:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '206':
          description: Partial content (range request)
          headers:
            Content-Range:
              schema:
                type: string
                example: bytes 0-1048575/5242880
            Content-Length:
              schema:
                type: integer
                format: int64
            Docker-Content-Digest:
              schema:
                type: string
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
        '404':
          $ref: '#/components/responses/BlobNotFound'
        '416':
          description: Range not satisfiable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    # end-10: Delete blob
    delete:
      operationId: deleteBlob
      tags: [blobs]
      summary: Delete blob
      description: |
        Delete a blob from the repository (OCI Distribution Spec end-10).
        Returns 202 Accepted (may be async garbage collection).
      responses:
        '202':
          description: Blob deletion accepted
        '404':
          $ref: '#/components/responses/BlobNotFound'
        '405':
          description: Method not allowed (blob is referenced by manifests)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  # end-4a: Start blob upload (monolithic or chunked)
...
components:
  parameters:
    RepositoryName:
      name: name
      in: path
      required: true
      schema:
        type: string
        pattern: '^[a-z0-9]+((\.|_|__|-+)[a-z0-9]+)*(\/[a-z0-9]+((\.|_|__|-+)[a-z0-9]+)*)*$'
        example: library/ubuntu
      description: Repository name (lowercase, alphanumeric with separators)
    Digest:
      name: digest
      in: path
      required: true
      schema:
        type: string
        pattern: '^(sha256|sha512):[a-f0-9]{64,128}$'
        example: sha256:abc123def456...
      description: Content digest (algorithm:hex)
...
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - errors
      properties:
        errors:
          type: array
          items:
            $ref: '#/components/schemas/Error'
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          enum:
            - BLOB_UNKNOWN
            - BLOB_UPLOAD_INVALID
            - BLOB_UPLOAD_UNKNOWN
            - DIGEST_INVALID
            - MANIFEST_BLOB_UNKNOWN
            - MANIFEST_INVALID
            - MANIFEST_UNKNOWN
            - NAME_INVALID
            - NAME_UNKNOWN
            - SIZE_INVALID
            - UNAUTHORIZED
            - DENIED
            - UNSUPPORTED
          description: OCI error code
        message:
          type: string
          description: Human-readable error message
        detail:
          type: object
          additionalProperties: true
          description: Additional error context
      example:
        code: BLOB_UNKNOWN
        message: Blob not found in repository
        detail:
          digest: sha256:abc123...
          repository: library/ubuntu
  responses:
...
    BlobNotFound:
      description: Blob not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            errors:
              - code: BLOB_UNKNOWN
                message: Blob not found in repository
...